+++ app/gui/src/presenter/project.rs
@@ -26,16 +26,17 @@ use model::project::VcsStatus;
 #[allow(unused)]
 #[derive(Debug)]
 struct Model {
-    controller:         controller::Project,
-    module_model:       model::Module,
-    graph_controller:   controller::ExecutedGraph,
-    ide_controller:     controller::Ide,
-    view:               view::project::View,
-    status_bar:         view::status_bar::View,
-    graph:              presenter::Graph,
-    code:               presenter::Code,
-    searcher:           RefCell<Option<presenter::Searcher>>,
-    available_projects: Rc<RefCell<Vec<ProjectMetadata>>>,
+    controller:           controller::Project,
+    module_model:         model::Module,
+    graph_controller:     controller::ExecutedGraph,
+    ide_controller:       controller::Ide,
+    view:                 view::project::View,
+    status_bar:           view::status_bar::View,
+    graph:                presenter::Graph,
+    code:                 presenter::Code,
+    searcher:             RefCell<Option<presenter::Searcher>>,
+    available_projects:   Rc<RefCell<Vec<ProjectMetadata>>>,
+    shortcut_transaction: RefCell<Option<Rc<model::undo_redo::Transaction>>>,
 }
 
 impl Model {
@@ -58,6 +59,7 @@ impl Model {
         let code = presenter::Code::new(text_controller, &view);
         let searcher = default();
         let available_projects = default();
+        let shortcut_transaction = default();
         Model {
             controller,
             module_model,
@@ -69,6 +71,7 @@ impl Model {
             code,
             searcher,
             available_projects,
+            shortcut_transaction,
         }
     }
 
@@ -175,6 +178,17 @@ impl Model {
         })
     }
 
+    /// Notification from shortcut manager that the handled shortcut has changed.
+    /// It is either the name of the command that was triggered or `None` when handling of the
+    /// last command was completed.
+    fn handled_shortcut_changed(&self, handled_shortcut: &Option<ImString>) {
+        debug!("Handled shortcut changed: {handled_shortcut:?}.");
+        let transaction = handled_shortcut
+            .as_ref()
+            .map(|shortcut| self.controller.model.urm().get_or_open_transaction(shortcut.as_str()));
+        *self.shortcut_transaction.borrow_mut() = transaction;
+    }
+
     fn toggle_component_browser_private_entries_visibility(&self) {
         let visibility = self.ide_controller.are_component_browser_private_entries_visible();
         self.ide_controller.set_component_browser_private_entries_visibility(!visibility);
@@ -208,6 +222,12 @@ impl Model {
     fn execution_finished(&self) {
         self.view.graph().frp.set_read_only(false);
         self.view.graph().frp.execution_finished.emit(());
+    }
+
+    fn execution_failed(&self) {
+        let message = crate::EXECUTION_FAILED_MESSAGE;
+        let message = view::status_bar::event::Label::from(message);
+        self.status_bar.add_event(message);
     }
 
     fn execution_context_interrupt(&self) {
@@ -360,6 +380,7 @@ impl Project {
             eval_ graph_view.execution_environment_play_button_pressed( model.trigger_clean_live_execution());
 
             eval_ view.go_to_dashboard_button_pressed (model.show_dashboard());
+            eval view.current_shortcut ((shortcut) model.handled_shortcut_changed(shortcut));
         }
 
         let graph_controller = self.model.graph_controller.clone_ref();
@@ -419,6 +440,9 @@ impl Project {
                 Notification::ExecutionFinished => {
                     model.execution_finished();
                 }
+                Notification::ExecutionFailed => {
+                    model.execution_failed();
+                }
             };
             std::future::ready(())
         });
@@ -464,7 +488,16 @@ impl Project {
         view: view::project::View,
         status_bar: view::status_bar::View,
     ) -> FallibleResult<Self> {
+        debug!("Initializing project controller...");
         let init_result = controller.initialize().await?;
-        Ok(Self::new(ide_controller, controller, init_result, view, status_bar))
+        debug!("Project controller initialized.");
+        let presenter =
+            Self::new(ide_controller, controller.clone(), init_result, view, status_bar);
+        debug!("Project presenter created.");
+        // Following the project initialization, the Undo/Redo stack should be empty.
+        // This makes sure that any initial modifications resulting from the GUI initialization
+        // won't clutter the undo stack.
+        controller.model.urm().repository.clear_all();
+        Ok(presenter)
     }
 }