+++ engine/runtime/src/main/java/org/enso/compiler/context/SimpleUpdate.java
@@ -1,0 +1,18 @@
+package org.enso.compiler.context;
+
+import org.enso.compiler.core.IR;
+import org.enso.text.editing.model.TextEdit;
+
+/**
+ * Simple editing change description.
+ *
+ * @param ir the current literal
+ * @param edit the editor change
+ * @param newIr the new literal
+ */
+public record SimpleUpdate(
+  IR.Literal ir,
+  TextEdit edit,
+  IR.Literal newIr
+) {
+}