+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/atom/GetFieldWithMatchNode.java
@@ -11,8 +11,8 @@ import org.enso.interpreter.runtime.callable.atom.Atom;
 import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
 import org.enso.interpreter.runtime.callable.atom.StructsLibrary;
 import org.enso.interpreter.runtime.callable.function.Function;
-import org.enso.interpreter.runtime.data.text.Text;
 import org.enso.interpreter.runtime.data.Type;
+import org.enso.interpreter.runtime.data.text.Text;
 import org.enso.interpreter.runtime.error.PanicException;
 
 @NodeInfo(shortName = "get_field", description = "A base for auto-generated Atom getters.")