+++ std-bits/table/src/main/java/org/enso/table/aggregations/GroupBy.java
@@ -1,9 +1,8 @@
 package org.enso.table.aggregations;
 
+import java.util.List;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.table.Column;
-
-import java.util.List;
 
 /** Aggregate Column getting the grouping key. */
 public class GroupBy extends Aggregator {